pyproject.toml
[tool.poetry]
name = "online-university"
version = "0.1.0"
description = ""
authors = ["glebokhlopkov <glebka_90@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
